+++ xknx/cemi/__init__.py
@@ -12,4 +12,10 @@ from .cemi_frame import (
 )
 from .cemi_handler import CEMIHandler
 from .const import CEMIErrorCode, CEMIMessageCode
-from .flags import CEMIFlags, CEMIFrameFormat, CEMIFrameType, CEMIPriority
+from .flags import (
+    CEMIAddressType,
+    CEMIFlags,
+    CEMIFrameFormat,
+    CEMIFrameType,
+    CEMIPriority,
+)