+++ xknx/secure/data_secure.py
@@ -211,7 +211,7 @@ class DataSecure:
                 key=key,
                 scf=s_apdu.scf,
                 address_fields_raw=_address_fields_raw,
-                dst_is_group_address=cemi_data.dst_is_group_address,
+                address_type=cemi_data.address_type,
                 tpci=cemi_data.tpci,
             )
         decrypted_payload = APCI.from_knx(plain_apdu_raw)
@@ -259,7 +259,7 @@ class DataSecure:
             sequence_number=self.get_sequence_number(),
             address_fields_raw=cemi_data.src_addr.to_knx()
             + cemi_data.dst_addr.to_knx(),
-            dst_is_group_address=cemi_data.dst_is_group_address,
+            address_type=cemi_data.address_type,
             tpci=cemi_data.tpci,
         )
         secure_cemi_data = copy(cemi_data)